Relocate RunSourceFactory
