+++ src/Services/Source/RunSourceFactory.php
@@ -2,26 +2,27 @@
 
 declare(strict_types=1);
 
-namespace App\Services;
+namespace App\Services\Source;
 
 use App\Entity\RunSource;
 use App\Entity\SourceOriginInterface;
 use App\Exception\EmptyEntityIdException;
-use App\Services\Source\Store;
+use App\Repository\SourceRepository;
+use App\Services\EntityIdFactory;
 use Symfony\Component\HttpFoundation\Request;
 
 class RunSourceFactory
 {
     public function __construct(
-        private readonly Store $store,
         private readonly EntityIdFactory $entityIdFactory,
+        private readonly SourceRepository $sourceRepository,
     ) {
     }
 
     /**
      * @throws EmptyEntityIdException
      */
-    public function createFromRequest(SourceOriginInterface $source, Request $request): RunSource
+    public function create(SourceOriginInterface $source, Request $request): RunSource
     {
         $parameters = [];
         foreach ($source->getRunParameterNames() as $runParameterName) {
@@ -31,7 +32,7 @@ class RunSourceFactory
         }
 
         $source = new RunSource($this->entityIdFactory->create(), $source, $parameters);
-        $this->store->add($source);
+        $this->sourceRepository->save($source);
 
         return $source;
     }
